+++ packages/breadcrumbs/src/styled/BreadcrumbsItem.jsx
@@ -1,14 +1,14 @@
 import styled from 'styled-components';
-import { akGridSizeUnitless, akFontSizeDefault } from '@atlaskit/util-shared-styles';
+import { gridSize, fontSize } from '@atlaskit/theme';
 
-const height = (akGridSizeUnitless * 3) / parseInt(akFontSizeDefault, 10);
+const height = (gridSize() * 3) / parseInt(fontSize(), 10);
 
 const BreadcrumbsItemElement = styled.div`
   display: flex;
   flex-direction: row;
   height: ${height}em;
   line-height: ${height}em;
-  margin: 0 ${akGridSizeUnitless / 2}px;
+  margin: 0 ${gridSize() / 2}px;
 `;
 
 BreadcrumbsItemElement.displayName = 'BreadcrumbsItemElement';