+++ packages/input/stories/00_single-line-text-input-story.jsx
@@ -1,6 +1,6 @@
 import { storiesOf, action } from '@kadira/storybook';
 import React from 'react';
-import SmartFieldBase from 'ak-field-base';
+import FieldBase from '@atlaskit/field-base';
 import SingleLineTextInput from '../src';
 import { name } from '../package.json';
 
@@ -69,7 +69,7 @@ storiesOf(storyName, module)
   ))
   .add('with field base', () => (
     <div style={containerStyle}>
-      <SmartFieldBase
+      <FieldBase
         label="Inside a FieldBase"
       >
         {createSingleLineTextInput({
@@ -77,6 +77,6 @@ storiesOf(storyName, module)
           autoFocus: true,
           isEditing: false,
         })}
-      </SmartFieldBase>
+      </FieldBase>
     </div>
   ));