+++ packages/navigation/test/unit/drawers/SearchDrawerSpec.jsx
@@ -3,35 +3,23 @@ import React from 'react';
 import { AkSearchDrawer } from '../../../src/index';
 import { searchIconOffset } from '../../../src/shared-variables';
 import Drawer from '../../../src/components/js/Drawer';
+import requiredProps from '../_drawer-util';
 
 describe('<SearchDrawer />', () => {
   describe('the inner Drawer', () => {
-    // icon for use in required icon props
-    const genericIcon = <span />;
-
     it('isFullWidth should pass width="full" to the inner drawer', () => {
       expect(mount(
-        <AkSearchDrawer
-          backIcon={genericIcon}
-          primaryIcon={genericIcon}
-          isFullWidth
-        />
+        <AkSearchDrawer {...requiredProps} isFullWidth />
       ).find(Drawer).props().width).toBe('full');
     });
     it('isFullWidth={false} should pass width="wide" to the inner drawer', () => {
       expect(mount(
-        <AkSearchDrawer
-          backIcon={genericIcon}
-          primaryIcon={genericIcon} isFullWidth={false}
-        />
+        <AkSearchDrawer {...requiredProps} isFullWidth={false} />
       ).find(Drawer).props().width).toBe('wide');
     });
     it('should render the backIcon in the correct position default to false', () => {
       expect(mount(
-        <AkSearchDrawer
-          backIcon={genericIcon}
-          primaryIcon={genericIcon}
-        />
+        <AkSearchDrawer {...requiredProps} />
       ).find(Drawer).props().backIconOffset).toBe(searchIconOffset);
     });
   });